+++ pakyow-core/lib/pakyow/logger/thread_local.rb
@@ -2,8 +2,6 @@
 
 require "forwardable"
 
-require "pakyow/support/deep_freeze"
-
 require "pakyow/logger"
 
 module Pakyow
@@ -11,12 +9,6 @@ module Pakyow
     # Determines at log time what logger to use, based on a thread-local context.
     #
     class ThreadLocal
-      extend Support::DeepFreeze
-      insulate :default
-
-      extend Forwardable
-      def_delegators :target, *(Logger.instance_methods - Object.instance_methods)
-
       def initialize(default_logger)
         @default = default_logger
       end
@@ -28,6 +20,21 @@ module Pakyow
       def replace(logger)
         @default = logger
       end
+
+      def silence(temporary_level = :error)
+        current = Thread.current[:pakyow_logger]
+
+        logger = target.dup
+        logger.level = temporary_level
+        Thread.current[:pakyow_logger] = logger
+
+        yield
+
+        Thread.current[:pakyow_logger] = current
+      end
+
+      extend Forwardable
+      def_delegators :target, *(Logger.instance_methods - instance_methods)
     end
   end
 end